Migrate GetInfo for expressions to cpp files
