+++ src/include/expression/case_expression.h
@@ -152,33 +152,9 @@ class CaseExpression : public AbstractExpression {
     }
   };
 
-  const std::string GetInfo(int num_indent) const override {
-    std::ostringstream os;
+  const std::string GetInfo(int num_indent) const override;
 
-    os << StringUtil::Indent(num_indent) << "Expression ::\n"
-       << StringUtil::Indent(num_indent + 1) << "expression type = Case,\n"
-       << StringUtil::Indent(num_indent + 1) << "default: \n"
-       << default_expr_.get()->GetInfo(num_indent + 2);
-
-    if (clauses_.size() > 0) {
-      os << StringUtil::Indent(num_indent + 1) << "clauses:\n";
-      for (const auto &clause : clauses_) {
-        os << StringUtil::Indent(num_indent + 2) << "first: \n"
-           << clause.first.get()->GetInfo(num_indent + 3);
-        os << StringUtil::Indent(num_indent + 2) << "second: \n"
-           << clause.second.get()->GetInfo(num_indent + 3);
-      }
-    }
-
-    return os.str();
-  }
-
-  const std::string GetInfo() const override {
-    std::ostringstream os;
-    os << GetInfo(0);
-
-    return os.str();
-  }
+  const std::string GetInfo() const override;
 
  private:
   // The list of case-when clauses
