+++ src/include/expression/constant_value_expression.h
@@ -66,24 +66,9 @@ class ConstantValueExpression : public AbstractExpression {
   // is null or not.
   bool IsNullable() const override { return value_.IsNull(); }
 
-  const std::string GetInfo(int num_indent) const override {
-    std::ostringstream os;
+  const std::string GetInfo(int num_indent) const override;
 
-    os << StringUtil::Indent(num_indent) << "Expression ::\n"
-       << StringUtil::Indent(num_indent + 1)
-       << "expression type = Constant Value,\n"
-       << StringUtil::Indent(num_indent + 1) << "value: " << value_.GetInfo()
-       << "\n";
-
-    return os.str();
-  }
-
-  const std::string GetInfo() const override {
-    std::ostringstream os;
-    os << GetInfo(0);
-
-    return os.str();
-  }
+  const std::string GetInfo() const override;
 
  private:
   // The constant
