+++ src/include/expression/function_expression.h
@@ -91,28 +91,9 @@ class FunctionExpression : public AbstractExpression {
     return func_arg_types_;
   }
 
-  const std::string GetInfo(int num_indent) const override {
-    std::ostringstream os;
+  const std::string GetInfo(int num_indent) const override;
 
-    os << StringUtil::Indent(num_indent) << "Expression ::\n"
-       << StringUtil::Indent(num_indent + 1) << "expression type = Function,\n"
-       << StringUtil::Indent(num_indent + 1) << "function name: " << func_name_
-       << "\n" << StringUtil::Indent(num_indent + 1) << "function args: "
-       << "\n";
-
-    for (const auto &child : children_) {
-      os << child.get()->GetInfo(num_indent + 2);
-    }
-
-    return os.str();
-  }
-
-  const std::string GetInfo() const override {
-    std::ostringstream os;
-    os << GetInfo(0);
-
-    return os.str();
-  }
+  const std::string GetInfo() const override;
 
  private:
   std::string func_name_;
